refactor(tests): consolidate DOM helpers into tests/helpers/helpers.ts
